Log all exception for the RefreshToken method in the AccountService

+++ Helpers/Utilities/Price.cs
@@ -7,7 +7,7 @@ namespace PEAS.Helpers.Utilities
         public const int MinPrice = 10_00;
         public const int MaxPrice = 5000_00;
 
-        public static string Format(int value)
+        public static string Format(long value)
         {
             double valueToDouble = value * 0.01;
             return String.Format("{0:0.00}", valueToDouble);
